requirements.txt
aiogram>=1.3.2
aiohttp>=3.3.2
aiosocksy>=0.1.1
arxiv>=0.2.2
beautifulsoup4>=4.6.0
facebook-sdk>=3.0.0
first>=2.0.1
langdetect>=1.0.7
multidict>=4.3.1
Pillow>=5.1.0
pytz>=2018.4
requests>=2.19.1
wikipedia>=1.4.0
peewee
Pillow
